services/synchronizer/requirements.txt
python-dotenv 
kafka-python
